+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [1.16.0] - 2018-07-16
+### Added
+- Redirect HTTP to HTTPS when "X-Forwarded-Proto" is used (Heroku or IBM Cloud (Bluemix))
+- Redirect HTTP to HTTPS in top page by specifying a command line option
+
+### Changed
+- Encrypt storing files by download-key from Basic Authentication
+- Specify HTTP/HTTPS ports by command line options
+
 ## [1.15.1] - 2018-07-13
 ### Added
 - Show file ID history on Web client
@@ -116,7 +125,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 - Adapt [`sbt-native-packager`](https://github.com/sbt/sbt-native-packager) for [Heroku](https://heroku.com/)
 
 
-[Unreleased]: https://github.com/nwtgck/trans-server-akka/compare/v1.15.1...HEAD
+[Unreleased]: https://github.com/nwtgck/trans-server-akka/compare/v1.16.0...HEAD
+[1.16.0]: https://github.com/nwtgck/trans-server-akka/compare/v1.15.1...v1.16.0
 [1.15.1]: https://github.com/nwtgck/trans-server-akka/compare/v1.15.0...v1.15.1
 [1.15.0]: https://github.com/nwtgck/trans-server-akka/compare/v1.14.0...v1.15.0
 [1.14.0]: https://github.com/nwtgck/trans-server-akka/compare/v1.13.0...v1.14.0